Контроллеры и модель для user и film исправление

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -4,13 +4,12 @@ import lombok.Data;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Positive;
-import java.time.Duration;
 import java.time.LocalDate;
 
 @Data
 public class Film {
 
-    int id;
+    long id;
     @NotEmpty
     String name;
     String description;
