+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -5,13 +5,12 @@ import lombok.Data;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Past;
-import javax.validation.constraints.Positive;
 import java.time.LocalDate;
 
 @Data
 public class User {
 
-    int id;
+    long id;
     @Email
     String email;
     @NotEmpty
